refact(CustomerListPage): break down components
